+++ backend/src/main/java/com/zjuse/bankingsystem/mapper/creditCard/CustomerMapper.java
@@ -15,7 +15,7 @@ public interface CustomerMapper {
     public List<CreditCard> queryCards(String id_number);
 
     @Update("insert into credit_card_application (id_number, credit_card_id, amount, type, status, password) values (#{id_number}, null, #{card_limit}, 1, 1, #{password})")
-    public void addNewCreditCard(String id_number, Double card_limit, String password);
+    public void addNewCreditCard(String id_number, BigInteger card_limit, String password);
 
     @Update("update credit_card set password = #{password} where id = #{card_id}")
     public void modifyCreditCardPassword(BigInteger card_id, String password);
